Merged PR 4959: Add new ACVP ECDSA error codes
This PR adds new ECDSA error codes to the ACVP testing backend.

----
#### AI description  (iteration 1)
#### PR Classification
New feature: Addition of new error codes for ECDSA handling.

#### PR Summary
This pull request introduces new error codes to enhance the ECDSA error handling capabilities in the backend. These changes are part of the integration of the acvpparser library with the firmware.
- `backend_ecdsa.h`: Added two new error codes, `BACKEND_ECDSA_INVALID_ECC_IMPLEMENTATION` and `BACKEND_ECDSA_API_TYPE_UNSUPPORTED`, to handle invalid ECC implementations and unsupported ECDSA API types.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #4902