Add support for RMA state transition through an authorized operation.
Rather than requiring RMA to be executed through the secure unlock flow, provide a mechanism to use the standard flow for any authorized operation.  This allows RMA handling to be more cleanly integrated with other functionality and makes it no different from other authorized commands.

The addition of authorized data handling to the authorization and execution flow made this type of integration possible.