Enhance FW update handlers to allow the automatic restore of the active image to be skipped.
During recovery flash boot scenarios, it may not always be desirable to restore the image on the active boot partition.  It's possible that the device was intentionally booted from the recovery partition, such as scenarios where an update is staged on flash waiting for an AC cycle.

Provide a mechanism for the firmware update handlers to be configured such that restoring the active image during recovery boot can be skipped.