Update host flash initialization manager to support static and const instances.
The handler for initializing the state of host SPI flash devices has been updated to work correctly with static and/or const SPI flash instances.  Rather than fully initializing the SPI flash instance, only the variable state is initialized by the handler.  The constant state is expected to be initialized statically or will be done as part of the handler's dynamic initialization flow.

As part of this update, the host flash initialization manager itself has been updated to support static and const instances.