Sanity check the signature length of a firmware component during verification.
The signature length field in the firmware component header is a 2-byte value, providing support for signatures much larger than would be used by ECC or RSA.  Since this is used prior to any verification to read in the signature data, it is easy for an attacker to force a large memory allocation and read operation, providing a way to stage unrelated data into memory.

This is mitigated in a couple ways:
1. Determine the maximum expected signature length and fail any components that have larger signatures without reading any of the signature data into memory.
2. Explicitly zeroize the signature memory after verification has completed.

Determining a reasonable value to use as a check against the signature length was not straightforward, since the firmware component verification has no visibility into what keys or algorithms are being used for verification.  To provide the necessary information, a new API was added to the signature verification abstraction to indicate what the maximum signature length is for the verification key being used.  In the ECC case, this also needed an additional API to get the maximum signature length based on a public key rather than the private key required by the existing API.