Add an extension builder for the standard Extended Key Usage extension.
By moving this to an extension builder, it allows the caller great flexibility for specifying any number of EKU OIDs to get added to a certificate.  There is no need to update any X.509 APIs or deal with internal library calls specific to this extension for custom OIDs.  This extension can be added to the cert by simply including the extension builder instance in the list provided when generating a certificate.

New files have been added to define well-known OIDs from different standards (DICE, SPDM, and X.509).  Not all OIDs that are defined in those specs have been included, but new ones are easy to add in the future as needed.

There has also been a little clean-up of the existing extension builders, such as adding missing tests and refactoring OIDs to align with the new handling of that data.

With this update, the EKU argument for CSR generation is redundant.  The ability to specify a single EKU is also insufficient for most current usage.  This argument has been marked as deprecated and should no longer be used.  It may get removed as part of a future change, but is left for now so that existing usage doesn't need to change.