Merged PR 5172: Remove Allowable Data sanity check that is out of spec with Cerberus Attestation.
Remove check that is out of spec with Project Cerberus Attestation Specification.
The attestation specification allows for a single Allowable Data element to contain data entries of different lengths, but the generator script is enforcing data elements to have the same length.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request removes an unnecessary data sanity check to align the code with Cerberus Attestation specification.

#### PR Summary
The changes eliminate the conditional check that compared the length of decoded data against a pre-existing value and raised an error on mismatch. Now, the data length is directly set based on the actual data, ensuring compliance with Cerberus Attestation.
- `tools/manifest_tools/manifest_parser.py`: Removed the conditional data length validation and error-raising logic, directly assigning the computed data length.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #5267