If host flash cannot be initialized, mark it as unsupported.
If a host SPI flash device cannot be fully initialized, do not spin indefinitely trying to get it to work.  Instead, attempt several retries and mark the flash as unsupported if initialization continues to fail.  This will ensure that the Cerberus and system boot sequence does not hang continuously retrying flash initialization.  Marking the flash as unsupported will ensure that no other host workflow can be executed.