Update the power on sequence to properly handle the case when the flash isn't supported.
If the flash configuration is unsupported (e.g. mismatched flash devices), the power on processing will fail and end up in bypass mode.  Since the error code returned when trying to apply the recovery image is not handled, the flash devices are swapped as part of the switch to bypass mode.  This behavior is undesirable.  Handle the unsupported flash scenario the same as when no recovery image could be applied.