Updates to ECC tests.
When using an alternate implementation for ECC with mbedtls, the external RNG provided to the sign operation cannot be deterministically tested since the usage by the alternate implementation is unknown.  In this configuration, skip these tests.

Also add test coverage for ECC signature verification with digests that are longer than the key length.