Add the framework to support handling fatal errors encountered during execution.
There are some scenarios during firmware execution that result in the system entering a bad state that is not recoverable without a device reset or some other extraordinary action.  However, there is no mechanism to easily trigger such an error in a generic way.  This creates a global handler, similar to the debug log, that can be leveraged by different components to signal that an unrecoverable error has occurred.  How (or if) the error handled is up to the system definition based on the instance assigned to the global reference.